COMMERCE-2500 mulesoft - move to appropriate package
